requirements.txt
--allow-external pyyaml
pyyaml

https://github.com/mitsuhiko/flask/tarball/master#egg=flask

# SQLAlchemy
Flask-SQLAlchemy

flask-restful